requirements.txt
ccxt
schedule
nest_asyncio
python-telegram-bot==20.7
